Week_8_labs/LAb - Agent_Deployment/requirements.txt
langgraph
langchain
langchain-openai
python-dotenv
streamlit 
pandas
plotly
regex